refactor: move ic module to mgmt
