+++ src/console/src/factory/mission_control.rs
@@ -7,7 +7,7 @@ use crate::types::state::MissionControl;
 use crate::wasm::mission_control_wasm_arg;
 use candid::Principal;
 use junobuild_shared::constants::CREATE_MISSION_CONTROL_CYCLES;
-use junobuild_shared::ic::create_canister_install_code;
+use junobuild_shared::mgmt::ic::create_canister_install_code;
 use junobuild_shared::types::state::UserId;
 
 pub async fn init_user_mission_control(
