+++ src/console/src/factory/satellite.rs
@@ -4,7 +4,7 @@ use crate::store::heap::{get_satellite_fee, increment_satellites_rate};
 use crate::wasm::satellite_wasm_arg;
 use candid::Principal;
 use junobuild_shared::constants::CREATE_SATELLITE_CYCLES;
-use junobuild_shared::ic::create_canister_install_code;
+use junobuild_shared::mgmt::ic::create_canister_install_code;
 use junobuild_shared::mgmt::cmc::cmc_create_canister_install_code;
 use junobuild_shared::mgmt::types::cmc::SubnetId;
 use junobuild_shared::types::interface::CreateCanisterArgs;
