+++ src/libs/shared/src/lib.rs
@@ -8,7 +8,6 @@ pub mod controllers;
 pub mod day;
 #[doc(hidden)]
 pub mod env;
-pub mod ic;
 #[doc(hidden)]
 pub mod impls;
 pub mod ledger;
