+++ src/libs/shared/src/mgmt/mod.rs
@@ -1,3 +1,4 @@
 pub mod cmc;
 mod impls;
 pub mod types;
+pub mod ic;
